# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Pharmaceutical Corporate Website API
  version: 1.0.0
  description: |
    REST API for the corporate website and admin CMS.

    **Conventions**
    - All responses use `{ success, message, data, meta? }`.
    - Public GET endpoints return only published/open/active records.
    - Admin endpoints require a Bearer JWT **or** the `access_token` HttpOnly cookie set by `/auth/login`.
    - Roles: `superadmin` (everything incl. users), `admin` (content + enquiries/applications/settings), `editor` (content & media only).
    - Content collections (`products`, `categories`, `capabilities`, `certifications`, `facilities`,
      `leadership`, `news`, `gallery`, `careers`) share the same CRUD shape described under *Generic content collection*.
servers:
  - url: http://localhost:5000/api/v1
...
security:
  - bearerAuth: []
  - cookieAuth: []
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
    cookieAuth: { type: apiKey, in: cookie, name: access_token }
  parameters:
    page: { name: page, in: query, schema: { type: integer, minimum: 1, default: 1 } }
    limit: { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100 } }
    search: { name: search, in: query, schema: { type: string } }
...
    id: { name: id, in: path, required: true, schema: { type: string, pattern: '^[0-9a-fA-F]{24}$' } }
...
  schemas:
    Envelope:
      type: object
      properties:
        success: { type: boolean }
        message: { type: string }
        data: {}
        meta:
          type: object
          properties:
            page: { type: integer }
            limit: { type: integer }
            total: { type: integer }
            totalPages: { type: integer }
...
  responses:
    OK: { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/Envelope' } } } }
    Created: { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/Envelope' } } } }
    NoContent: { description: Deleted }
...
paths:
...
  /admin/media/upload:
    post:
      tags: [Admin]
      summary: Upload up to 5 public files (JPG/PNG/WebP/PDF). Content roles.
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                files: { type: array, items: { type: string, format: binary } }
                folder: { type: string, enum: [products, categories, gallery, news, certificates, facilities, leadership, pages, documents, misc] }
                alt: { type: string }
      responses: { '201': { $ref: '#/components/responses/Created' } }
  /admin/media:
    get:
      tags: [Admin]
      parameters: [{ $ref: '#/components/parameters/page' }, { $ref: '#/components/parameters/limit' }, { name: kind, in: query, schema: { type: string, enum: [image, document] } }, { name: folder, in: query, schema: { type: string } }, { $ref: '#/components/parameters/search' }]
      responses: { '200': { $ref: '#/components/responses/OK' } }
  /admin/media/{id}:
    patch: { tags: [Admin], summary: Update alt text, parameters: [{ $ref: '#/components/parameters/id' }], requestBody: { content: { application/json: { schema: { type: object, properties: { alt: { type: string } } } } } }, responses: { '200': { $ref: '#/components/responses/OK' } } }
    delete: { tags: [Admin], parameters: [{ $ref: '#/components/parameters/id' }], responses: { '204': { $ref: '#/components/responses/NoContent' } } }
  /admin/users:
    get: { tags: [Admin], summary: List admin users (superadmin), responses: { '200': { $ref: '#/components/responses/OK' } } }
    post:
      tags: [Admin]
      summary: Create admin/editor (superadmin)
      requestBody: { content: { application/json: { schema: { type: object, required: [name, email, password, role], properties: { name: { type: string }, email: { type: string }, password: { type: string }, role: { type: string, enum: [superadmin, admin, editor] } } } } } }
      responses: { '201': { $ref: '#/components/responses/Created' } }
  /admin/users/{id}:
    put: { tags: [Admin], parameters: [{ $ref: '#/components/parameters/id' }], requestBody: { content: { application/json: { schema: { type: object, properties: { name: { type: string }, role: { type: string }, isActive: { type: boolean }, password: { type: string } } } } } }, responses: { '200': { $ref: '#/components/responses/OK' } } }
    delete: { tags: [Admin], parameters: [{ $ref: '#/components/parameters/id' }], responses: { '204': { $ref: '#/components/responses/NoContent' } } }
...